fix : bug fixxed and console.log() removed #11
